+++ osu.Game/Online/Multiplayer/GameTypes/GameTypeVersus.cs
@@ -1,3 +1,6 @@
+// Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
+// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
+
 using osu.Framework.Graphics;
 using osu.Game.Graphics;
