+++ osu.Game/Screens/Multi/Match/Components/MatchLeaderboard.cs
@@ -4,13 +4,11 @@
 using System;
 using System.Collections.Generic;
 using osu.Framework.Allocation;
-using osu.Game.Graphics;
 using osu.Game.Online.API;
 using osu.Game.Online.API.Requests;
 using osu.Game.Online.API.Requests.Responses;
 using osu.Game.Online.Leaderboards;
 using osu.Game.Online.Multiplayer;
-using osu.Game.Scoring;
 
 namespace osu.Game.Screens.Multi.Match.Components
 {
@@ -55,27 +53,6 @@ namespace osu.Game.Screens.Multi.Match.Components
         }
 
         protected override LeaderboardScore CreateDrawableScore(APIRoomScoreInfo model, int index) => new MatchLeaderboardScore(model, index);
-    }
-
-    public class MatchLeaderboardScore : LeaderboardScore
-    {
-        public MatchLeaderboardScore(APIRoomScoreInfo score, int rank)
-            : base(score, rank)
-        {
-        }
-
-        [BackgroundDependencyLoader]
-        private void load()
-        {
-            RankContainer.Alpha = 0;
-        }
-
-        protected override IEnumerable<LeaderboardScoreStatistic> GetStatistics(ScoreInfo model) => new[]
-        {
-            new LeaderboardScoreStatistic(FontAwesome.fa_crosshairs, "Accuracy", string.Format(model.Accuracy % 1 == 0 ? @"{0:P0}" : @"{0:P2}", model.Accuracy)),
-            new LeaderboardScoreStatistic(FontAwesome.fa_refresh, "Total Attempts", ((APIRoomScoreInfo)model).TotalAttempts.ToString()),
-            new LeaderboardScoreStatistic(FontAwesome.fa_check, "Completed Beatmaps", ((APIRoomScoreInfo)model).CompletedBeatmaps.ToString()),
-        };
     }
 
     public enum MatchLeaderboardScope
